\name{dataSets}
\alias{dists}
\alias{eps}
\alias{eqtl}
\alias{gmap}
\alias{maps}
\alias{markers}
\alias{gdat}
\alias{mdat}
\alias{mpos}
\alias{qtl}
\alias{traits}
\alias{v}
\alias{xid}
\docType{data}
\title{
  Data sets
}
\description{
  A collection of data sets for demonstration.
}
\usage{
data(etrait)
}
\keyword{datasets}

