\name{mlogLik}
\alias{mlogLik}
\title{
Extract log-likelihood
}
\description{
Extract log-likelihood from the object of a multivariate multiple regression model.
}
\usage{
mlogLik(object)
}
\arguments{
  \item{object}{
object of a fitted multivariate multiple regression model.
}
}

\value{
log-likelihood.
}

